Generic ClientIdAuthenticator
